requirements.txt
django==2.2.2
qrcode==6.1 
Pillow==6.2.1 